backend/pyproject.toml
[project]
name = "ai-hr-automator"
version = "0.1.0"
description = "A bare-minimum AI email automation tool for job seekers"
readme = "README.md"
requires-python = ">=3.10"
dependencies = [
    "fastapi>=0.109.0",
    "uvicorn[standard]>=0.27.0",
    "pydantic-settings>=2.1.0",
    "pydantic[email]>=2.5.0",
    "google-generativeai>=0.3.0",
    "python-multipart>=0.0.7",
]

[build-system]
requires = ["setuptools>=61.0"]
build-backend = "setuptools.build_meta"